sdkconfig: update for debugging
Change the baud rate for both the monitor and UART to 115200.

Also update the IP timeout and max stack size. Doesn't seem to help
but doesn't hurt either. I get responses from time to time that the
headers are too long for the server to process. I suspect Chrome is
sending more than it absolutely needs to but didn't really dig into
this too carefully yet.

Signed-off-by: Joel Nider <joel@nider.org>